src/requirements.txt
# This file was generated by 'pip freeze' command in order to fixate state of the system
# this was done because sometimes, after recompilation of the docker container we run into the situation that
# system stops working, and we can't spend a lot of the time to investigate the problem, so if you need to update system than:
# 1. Delete ==<version>" prefixes
# 2. Leave only general libraries
# 3. Recompile the docker container
# 4. Insure that all working as it should
# 5. Generate list of the packages by 'pip freeze' command
# 6. Update requirements.txt
# ======================================================================================
# List of general libs that are required by this service
psycopg2==2.6.1
requests==2.9.1
requests-oauthlib==0.6.1
pp-ez==0.2.0
pprintpp==0.2.3
Django==1.9.5
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
mock==2.0.0

# ======================================================================================
# All other libraries that was installed during general libs installation
aiohttp==0.21.5
asyncio==3.4.3
chardet==2.3.0
defusedxml==0.4.1
oauthlib==1.0.3
pbr==1.8.1
PyJWT==1.4.0
six==1.10.0